Refactor APIs to work on Floats instead of Ints.
This applies to the Canvas object, the Scene API, the Viewport, and the
geometry. The reasoning is that these are used for
representing/rendering world objects, which should not be in pixel
coordinates, because they can't (due to physics simulation and
precision).

Canvas can, on top of that, be scaled and rotated, which will make its
integer-based API insufficient. The commit contains an example in the
hello project, which shows what can be done now that would have not
worked very well before (leading to square of uneven sizes).

+++ core/src/main/scala/sgl/WindowProvider.scala
@@ -82,7 +82,7 @@ trait WindowProvider {
     /** Convert an amount of DIP to the same amount of pixel. */
     def dp2px(x: Int): Int = (x*Window.ppi/160f).toInt
     /** Convert an amount of DIP to the same amount of pixel. */
-    def dp2px(x: Double): Double = x*Window.ppi/160f
+    def dp2px(x: Float): Float = x*Window.ppi/160f
   }
 
   /** the unique window hosting the game
