+++ core/src/main/scala/sgl/geometry/Point.scala
@@ -3,12 +3,12 @@ package sgl.geometry
 import scala.language.implicitConversions
 
 /*
- * Point and Vec are doubles as they are used to simulate physics
+ * Point and Vec are floats as they are used to simulate physics
  * and using Int (pixels) tend to lose precision when a frame only
- * move modify a fraction of a pixel
+ * move a fraction of a pixel.
  */
 
-case class Point(x: Double, y: Double) {
+case class Point(x: Float, y: Float) {
 
   def +(m: Vec): Point = Point(x+m.x, y+m.y)
   def -(m: Vec): Point = Point(x-m.x, y-m.y)
@@ -24,6 +24,6 @@ case class Point(x: Double, y: Double) {
 }
 
 object Point {
-  implicit def tupleToPoint(p: (Double, Double)): Point = Point(p._1, p._2)
+  implicit def tupleToPoint(p: (Float, Float)): Point = Point(p._1, p._2)
   implicit def intTupleToPoint(p: (Int, Int)): Point = Point(p._1, p._2)
 }
