+++ core/src/main/scala/sgl/geometry/Vec.scala
@@ -7,16 +7,16 @@ package sgl.geometry
  * they can be seen as the same object, because it seems
  * like we don't want to accidently replace one by another
  */
-case class Vec(x: Double, y: Double) {
+case class Vec(x: Float, y: Float) {
 
   def +(m: Vec): Vec = Vec(x+m.x, y+m.y)
   def -(m: Vec): Vec = Vec(x-m.x, y-m.y)
 
-  def *(s: Double): Vec = Vec(x*s, y*s)
+  def *(s: Float): Vec = Vec(x*s, y*s)
   
   def unary_- : Vec = Vec(-x, -y)
 
-  def norm: Double = math.sqrt(x*x + y*y)
+  def norm: Float = math.sqrt(x*x + y*y).toFloat
 
   def normalized: Vec = {
     val n = this.norm
@@ -29,8 +29,8 @@ case class Vec(x: Double, y: Double) {
   def pmax(that: Vec): Vec = Vec(x max that.x, y max that.y)
   def pmin(that: Vec): Vec = Vec(x min that.x, y min that.y)
 
-  def dotProduct(that: Vec): Double = this.x*that.x + this.y*that.y
-  def *(that: Vec): Double = this.dotProduct(that)
+  def dotProduct(that: Vec): Float = this.x*that.x + this.y*that.y
+  def *(that: Vec): Float = this.dotProduct(that)
 
   //TODO: could define +=, *=, -= as mutating the Vec? would make much
   //      code similar, we would still keep +/*/- as operation that
